support nested field paths in http paths
