+++ lib/grpc/server/transcode/query.ex
@@ -1,6 +1,5 @@
 defmodule GRPC.Server.Transcode.Query do
   # This module is based on https://github.com/elixir-plug/plug/blob/main/lib/plug/conn/query.ex
-
   @moduledoc """
   Decoding of URL-encoded queries as per the rules outlined in the documentation for [`google.api.HttpRule`](https://cloud.google.com/endpoints/docs/grpc-service-config/reference/rpc/google.api#google.api.HttpRule)
 
@@ -10,6 +9,8 @@ defmodule GRPC.Server.Transcode.Query do
   1. Sub-paths on the form `path.subpath` are decoded as nested maps
   1. Sub-paths with the same leaf key are decoded as a list
   """
+
+  alias GRPC.Server.Transcode.FieldPath
 
   @doc """
   Decodes the given `query`.
@@ -64,26 +65,10 @@ defmodule GRPC.Server.Transcode.Query do
           {decode_www_form(key), ""}
       end
 
-    decode_pair(current, acc)
+    FieldPath.decode_pair(current, acc)
   end
 
   defp decode_www_form(value) do
     URI.decode_www_form(value)
-  end
-
-  defp decode_pair({key, value}, acc) do
-    parts = :binary.split(key, ".", [:global])
-    assign_map(parts, value, acc)
-  end
-
-  defp assign_map(parts, value, acc) do
-    {_, acc} =
-      get_and_update_in(acc, Enum.map(parts, &Access.key(&1, %{})), fn
-        prev when prev == %{} -> {prev, value}
-        prev when is_list(prev) -> {prev, [value | prev]}
-        prev -> {prev, [value, prev]}
-      end)
-
-    acc
   end
 end
